Add tools and refactor the main file
