+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/package.json
@@ -1,5 +1,18 @@
 {
   "name": "w02e04-dungeons-and-the-princess",
   "version": "1.0.0",
-  "dependencies": {}
+  "scripts": {
+    "dev": "nodemon src/main.ts",
+    "build": "tsc",
+    "build:vite": "vite build",
+    "test": "jest"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.3",
+    "vitest": "^0.24.3",
+    "vite": "^2.9.15",
+    "vite-plugin-node": "^1.0.0"
+  }
 }
