+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/model/castle.ts
@@ -1,0 +1,11 @@
+import { Dungeon } from './dungeon';
+import { Person } from './person';
+
+export class Castle {
+    constructor(private dungeon: Dungeon) { }
+
+    gotoTheDungeon(daredevil: Person) {
+        // #RQ1: peasant cannot be at the Castle !
+        this.dungeon.enter(daredevil)
+    }
+}
