+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/model/dungeon.ts
@@ -1,0 +1,11 @@
+import { Person } from './person';
+import { Underground } from './underground';
+
+export class Dungeon {
+    constructor(private underground: Underground) { }
+
+    enter(p: Person) {
+        // #RQ2: queen cannot enter the dungeon!
+        this.underground.enter(p)
+    }
+}
