+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/model/person.ts
@@ -1,0 +1,13 @@
+export type Title = 'princess' | 'king' | 'queen' | 'knight' | 'peasant'
+
+export class Person {
+    constructor(private name: string, public readonly title: Title) { }
+
+    shoutTheName() {
+        console.log(`I'am ${this.name}! The ${this.title}.`)
+    }
+
+    sayThanks() {
+        console.log(`${this.name}: Thank you...`)
+    }
+}
