+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/model/underground.ts
@@ -1,0 +1,36 @@
+import { Person } from './person'
+
+export class Underground {
+    private static INITIAL_NO_OF_BARRICADES = 3
+    private noOfBarricades = Underground.INITIAL_NO_OF_BARRICADES
+
+    constructor(private prisoner: Person) { }
+
+    enter(savior: Person) {
+        // #RQ3: king cannot enter Underground !
+        savior.shoutTheName()
+        this.saveThePrisoner(savior)
+    }
+
+    private saveThePrisoner(savior: Person) {
+        // #RQ4: only knight can attempt to save prisoner!
+        // #RQ5: If not knight Evil dragon will set up new barricade
+        // #RQ6: one knight will remove one barricade
+        if (this.hasNoBarricades()) {
+            this.prisoner.sayThanks()
+        }
+    }
+
+    // Clean state:
+    initTheBarricades() {
+        this.noOfBarricades = Underground.INITIAL_NO_OF_BARRICADES
+    }
+
+    private removeBarricade() {
+        this.noOfBarricades -= 1
+    }
+
+    private hasNoBarricades() {
+        return this.noOfBarricades === 0
+    }
+}
