+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+      "target": "ES2021",
+      "module": "commonjs",
+      "outDir": "./dist",
+      "rootDir": "./src",
+      "allowJs": false,
+      "esModuleInterop": true,
+      "forceConsistentCasingInFileNames": true,
+      "strict": true,
+      "skipLibCheck": true
+    },
+    "exclude": ["./**/*.spec.ts"]
+  }
