+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite';
+import { VitePluginNode } from 'vite-plugin-node';
+export default defineConfig({
+    plugins: [
+        ...VitePluginNode({
+            adapter: 'nest',
+            appPath: './src/main.ts',
+            exportName: 'viteNodeApp',
+            tsCompiler: 'esbuild',
+        })
+    ],
+});
